fix console routes
