+++ src/routes/consoles.ts
@@ -19,25 +19,6 @@ export default function () {
       } catch (error) {
         next(error);
       }
-    },
-    
-    // TODO: Move PUT and DELETE to src/routes/consoles/{consoleName}.ts
-    async PUT (request: Request, response: Response, next: NextFunction) {
-      try {
-        const console = await ConsoleController.putConsole(request.params.consoleName, request.body);
-        response.status(200).send(console);
-      } catch (error) {
-        next(error);
-      }
-    },
-
-    async DELETE (request: Request, response: Response, next: NextFunction) {
-      try {
-        const console = await ConsoleController.deleteConsole(request.params.consoleName);
-        response.status(200).send(console);
-      } catch (error) {
-        next(error);
-      }
     }
   };
 }
