+++ src/server.ts
@@ -2,7 +2,7 @@ import express, { Application, Request, Response } from 'express';
 import BodyParser from 'body-parser';
 import { initialize } from 'express-openapi';
 import yaml from 'yamljs';
-import { readFileSync, writeFileSync } from 'fs';
+import { readFileSync } from 'fs';
 import path, { resolve } from 'path';
 import { resolveRefsAt } from 'json-refs';
 import swaggerUi from 'swagger-ui-express';
